moved more classes into packages
