--- battleship/src/battleship/Shot.java
@@ -1,26 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package battleship;
-
-/**
- *
- * @author shatziwebster
- */
-public class Shot {
-    
-    
-        public Shot ()
-    {
-        boolean shot;
-       if (shot=true){
-           System.out.println("You have made a strike");
-       }                 
-       else {System.out.println("You have missed");
-    }
-    
-    }
-    
-}
