+++ battleship/src/citbyui260/section03/battleship/ships/ShipType.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package battleship;
+package citbyui260.section03.battleship.ships;
 
 /**
  *
